sqlite sequences do not need to be micromanaged

+++ _test/SchemaBuilder.test.php
@@ -40,6 +40,7 @@ class schemaBuilder_struct_test extends \DokuWikiTest {
 
         foreach ($tableNames as $tableName) {
             if ($tableName == 'opts') continue;
+            if (substr($tableName, 0, 7) == 'sqlite_') continue;
             $this->sqlite->query('DROP TABLE ?', $tableName);
         }
 
@@ -47,7 +48,6 @@ class schemaBuilder_struct_test extends \DokuWikiTest {
         $this->sqlite->query("CREATE TABLE schema_assignments ( assign NOT NULL, tbl NOT NULL, PRIMARY KEY(assign, tbl) );");
         $this->sqlite->query("CREATE TABLE schema_cols ( sid INTEGER REFERENCES schemas (id), colref INTEGER NOT NULL, enabled BOOLEAN DEFAULT 1, tid INTEGER REFERENCES types (id), sort INTEGER NOT NULL, PRIMARY KEY ( sid, colref) )");
         $this->sqlite->query("CREATE TABLE schemas ( id INTEGER PRIMARY KEY AUTOINCREMENT, tbl NOT NULL, ts INT NOT NULL, chksum DEFAULT '' )");
-        $this->sqlite->query("CREATE TABLE sqlite_sequence(name,seq)");
         $this->sqlite->query("CREATE TABLE types ( id INTEGER PRIMARY KEY AUTOINCREMENT, class NOT NULL, ismulti BOOLEAN DEFAULT 0, label DEFAULT '', config DEFAULT '' )");
         $this->sqlite->query("CREATE TABLE multivals ( tbl NOT NULL, colref INTEGER NOT NULL, pid NOT NULL, rev INTEGER NOT NULL, row INTEGER NOT NULL, value, PRIMARY KEY(tbl, colref, pid, rev, row) )");
     }
